+++ src/Reserve-Copy-Pdf.ps1
@@ -1,0 +1,12 @@
+param(
+    [string]$FileName = "NewtonsInterpolationFormulaAndSumsOfPowers"
+)
+
+$RootDirectory = "$PSScriptRoot/.."
+
+cd $RootDirectory
+
+$SourcePath = "$RootDirectory/out/$FileName.pdf"
+$DestinationPath = "$RootDirectory/out/$FileName.annotated.pdf"
+
+Copy-Item -Path $SourcePath -Destination $DestinationPath -Force