+++ CHANGELOG.md
@@ -11,7 +11,20 @@ and this project adheres to [Semantic Versioning v2.0.0](https://semver.org/spec
 
 [//]: # (### Changed)
 
+## [1.6.0] - In progress
+
+### Changed
+
+- Update pandoc script
+- Update pandoc metadata
+- Add math writing
+- Vs code build configs
+- Updates in README
+- Updates in footer
+
 ## [1.5.0] - 06 Jan 2026
+
+### Changed
 
 - Reuse build template actions
 - Add notation for falling and rising factorials
@@ -20,8 +33,6 @@ and this project adheres to [Semantic Versioning v2.0.0](https://semver.org/spec
 - In bibliography: note -> howpublished
 - Update init script
 - Implement HTML deploy to GitHub pages
-
-### Changed
 
 ## [1.4.0] - 06 Jan 2026